requirements.txt
pandas 
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
ensure
joblib
types-PyYAML
Flask
Flask-Cors
-e .